update and rename test code
